+++ src/js/services/scene/camera.js
@@ -5,36 +5,18 @@ const THREE = require('three');
 window.THREE = window.THREE || THREE;
 
 const degreeToRadiant = deg => Math.PI / (180 / deg);
-const calcucalateAngle = (viewport, range) => ({
-	x: (viewport.mouse.x / viewport.screen.width * range.h) + range.hOffset,
-	y: (viewport.mouse.y / viewport.screen.height * range.v) + range.vOffset
-});
 
 const init = () => {};
 
-let cameraAngle = {x: 45, y: 210};
-let mouse = false;
-
 const refresh = ({camera, state}) => {
+	// Camera angle is now calculated in services/control/camera.js
+	const cameraAngle = state.camera.angle || {x: 45, y: 210};
+	
 	const centerPos = new THREE.Vector3().fromArray(
 		state.camera.followPlayer
 			? state.player.position
 			: [0, 0, 0]
 	);
-
-	if (state.viewport.mouse.down) {
-		cameraAngle = {
-			x: (360 + (state.player.rotation - 90 - 45) % 360),
-			y: (state.camera.range.h +
-				cameraAngle.y - (mouse ? (mouse.y - state.viewport.mouse.y) * 0.3 : 0)) % state.camera.range.h
-		};
-		// console.log(cameraAngle, mouse);
-		cameraAngle.y = Math.max(Math.min(cameraAngle.y, 260), 170);
-
-		mouse = {...state.viewport.mouse};
-	} else {
-		mouse = false;
-	}
 
 	const minDistance = 12;
 	const distance = Math.min(
@@ -44,7 +26,6 @@ const refresh = ({camera, state}) => {
 	const lookAtPos = centerPos.clone().add({
 		x: 0, y: 6 - 6 * ((cameraAngle.y - 180) / 90), z: 0
 	});
-	// console.log(cameraAngle.y);
 
 	camera.position.copy(lookAtPos.clone().add({
 		// x